Convert RoleSeeder and LLMSeeder to use enums.

+++ database/seeders/StaticDataSeeders/EnumSeeder.php
@@ -2,22 +2,28 @@
 
 namespace Database\Seeders\StaticDataSeeders;
 
+use App\Enums\Contracts\SeedableEnum;
 use Illuminate\Database\Seeder;
 
-abstract class StaticDataSeeder extends Seeder
+abstract class EnumSeeder extends Seeder
 {
-    protected static $staticDataClass;
-    protected static $forceDelete = false;
+    protected string $enum;
+    protected static bool $forceDelete = false;
 
     public function run(): void
     {
-        $staticData = static::$staticDataClass::getData();
-        $modelClass = static::$staticDataClass::getModelClass();
+        if (!is_subclass_of($this->enum, SeedableEnum::class)) {
+            throw new \InvalidArgumentException("Enum must implement SeedableEnum");
+        }
+
+        $staticData = $this->enum::getSeedData();
+        $modelClass = $this->enum::getModelClass();
 
         $model = new $modelClass;
         $modelClass::upsert($staticData, [$model->getKeyName()]);
 
         $deleteQuery = $modelClass::whereNotIn($model->getKeyName(), collect($staticData)->pluck($model->getKeyName()));
+
         if (static::$forceDelete) {
             $deleteQuery->forceDelete();
         } else {
